+++ app/controllers/mattermost/project_settings_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Mattermost
-  class ProjectSettingsController < ApplicationController
+  class ProjectSettingsController < ::ApplicationController
     before_action :find_project_by_project_id
     before_action :authorize